ML/chatbot/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
flask-restx==1.3.0
python-dotenv==1.1.0
google-cloud-aiplatform==1.73.0
